+++ packages/@ember/-internals/glimmer/lib/helpers/action.ts
@@ -2,7 +2,7 @@
 @module ember
 */
 import { get } from '@ember/-internals/metal';
-import type { AnyFn } from '@ember/-internals/utils/types';
+import type { AnyFn } from '@ember/-internals/utility-types';
 import { assert } from '@ember/debug';
 import { flaggedInstrument } from '@ember/instrumentation';
 import { join } from '@ember/runloop';
