+++ packages/@ember/-internals/metal/lib/events.ts
@@ -4,7 +4,7 @@
 import type { Meta } from '@ember/-internals/meta';
 import { meta as metaFor, peekMeta } from '@ember/-internals/meta';
 import { setListeners } from '@ember/-internals/utils';
-import type { AnyFn } from '@ember/-internals/utils/types';
+import type { AnyFn } from '@ember/-internals/utility-types';
 import { assert } from '@ember/debug';
 
 /*
