+++ packages/@ember/routing/route.ts
@@ -14,7 +14,7 @@ import { A as emberA } from '@ember/array';
 import { ActionHandler } from '@ember/-internals/runtime';
 import { isEmpty, typeOf } from '@ember/utils';
 import { isProxy, lookupDescriptor } from '@ember/-internals/utils';
-import type { AnyFn } from '@ember/-internals/utils/types';
+import type { AnyFn } from '@ember/-internals/utility-types';
 import Controller from '@ember/controller';
 import type { ControllerQueryParamType } from '@ember/controller';
 import { assert, info, isTesting } from '@ember/debug';
