+++ packages/@ember/routing/router.ts
@@ -46,7 +46,7 @@ import Router, { logAbort, STATE_SYMBOL } from 'router_js';
 import type { Timer } from 'backburner';
 import EngineInstance from '@ember/engine/instance';
 import type { QueryParams } from 'route-recognizer';
-import type { AnyFn, MethodNamesOf, OmitFirst } from '@ember/-internals/utils/types';
+import type { AnyFn, MethodNamesOf, OmitFirst } from '@ember/-internals/utility-types';
 import type { Template } from '@glimmer/interfaces';
 import type ApplicationInstance from '@ember/application/instance';
