+++ packages/@ember/runloop/index.ts
@@ -3,7 +3,7 @@ import { onErrorTarget } from '@ember/-internals/error-handling';
 import { flushAsyncObservers } from '@ember/-internals/metal';
 import type { DeferredActionQueues } from 'backburner';
 import Backburner, { Timer } from 'backburner';
-import type { AnyFn } from '@ember/-internals/utils/types';
+import type { AnyFn } from '@ember/-internals/utility-types';
 
 export { Timer };
